docs: fix cta and alert shortcodes
refactors and fixes the cta and alert shortcodes
removes the `width` attr from example shortcode
